pyproject.toml
[tool.poetry]
name = "senamhi-tracker"
version = "0.1.0"
description = "Web scraper and monitor for SENAMHI Peru weather data"
authors = ["rosepb28 <rosamariapb28@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.5"
beautifulsoup4 = "^4.14.2"
lxml = "^6.0.2"
sqlalchemy = "^2.0.44"
alembic = "^1.17.1"
typer = "^0.20.0"
python-dateutil = "^2.9.0"
pydantic-settings = "^2.12.0"
python-dotenv = "^1.2.1"
rich = "^13.9.4"
schedule = "^1.2.2"
python-daemon = "^3.1.2"
flask = "^3.1.2"
openmeteo-requests = "^1.7.4"
requests-cache = "^1.2.1"
retry-requests = "^2.0.0"
pandas = "^2.3.3"
pyyaml = "^6.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.1"
pytest-cov = "^7.0.0"
responses = "^0.25.8"
ruff = "^0.14.4"
pre-commit = "^4.4.0"

[tool.poetry.scripts]
senamhi = "app.main:app"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
